+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -55,7 +55,6 @@ import com.palantir.atlasdb.config.AtlasDbRuntimeConfig;
 import com.palantir.atlasdb.config.ImmutableAtlasDbConfig;
 import com.palantir.atlasdb.config.ImmutableLeaderRuntimeConfig;
 import com.palantir.atlasdb.config.ImmutableServerListConfig;
-import com.palantir.atlasdb.config.ImmutableTimeLockClientConfig;
 import com.palantir.atlasdb.config.LeaderConfig;
 import com.palantir.atlasdb.config.LeaderRuntimeConfig;
 import com.palantir.atlasdb.config.ServerListConfig;
@@ -63,7 +62,6 @@ import com.palantir.atlasdb.config.ServerListConfigs;
 import com.palantir.atlasdb.config.SweepConfig;
 import com.palantir.atlasdb.config.TargetedSweepInstallConfig;
 import com.palantir.atlasdb.config.TargetedSweepRuntimeConfig;
-import com.palantir.atlasdb.config.TimeLockClientConfig;
 import com.palantir.atlasdb.config.TimestampClientConfig;
 import com.palantir.atlasdb.factory.Leaders.LocalPaxosServices;
 import com.palantir.atlasdb.factory.startup.ConsistencyCheckRunner;
@@ -146,7 +144,6 @@ import com.palantir.timestamp.TimestampStoreInvalidator;
 import com.palantir.tritium.metrics.registry.DefaultTaggedMetricRegistry;
 import com.palantir.tritium.metrics.registry.TaggedMetricRegistry;
 import com.palantir.util.JavaSuppliers;
-import com.palantir.util.OptionalResolver;
 
 @Value.Immutable
 @Value.Style(stagedBuilder = true)
@@ -349,7 +346,7 @@ public abstract class TransactionManagers {
                 closeables);
 
         Callback<TransactionManager> callbacks = new Callback.CallChain<>(ImmutableList.of(
-                timelockConsistencyCheckCallback(config, runtimeConfigSupplier.get(), lockAndTimestampServices),
+                timelockConsistencyCheckCallback(runtimeConfigSupplier.get(), lockAndTimestampServices),
                 targetedSweep.singleAttemptCallback(),
                 asyncInitializationCallback()));
 
@@ -589,10 +586,9 @@ public abstract class TransactionManagers {
     }
 
     private static Callback<TransactionManager> timelockConsistencyCheckCallback(
-            AtlasDbConfig atlasDbConfig,
             AtlasDbRuntimeConfig initialRuntimeConfig,
             LockAndTimestampServices lockAndTimestampServices) {
-        if (isUsingTimeLock(atlasDbConfig, initialRuntimeConfig)) {
+        if (isUsingTimeLock(initialRuntimeConfig)) {
             // Only do the consistency check if we're using TimeLock.
             // This avoids a bootstrapping problem with leader-block services without async initialisation,
             // where you need a working timestamp service to check consistency, you need to check consistency
@@ -607,8 +603,8 @@ public abstract class TransactionManagers {
         return Callback.noOp();
     }
 
-    private static boolean isUsingTimeLock(AtlasDbConfig atlasDbConfig, AtlasDbRuntimeConfig runtimeConfig) {
-        return atlasDbConfig.timelock().isPresent() || runtimeConfig.timelockRuntime().isPresent();
+    private static boolean isUsingTimeLock(AtlasDbRuntimeConfig runtimeConfig) {
+        return runtimeConfig.timelockRuntime().isPresent();
     }
 
     /**
@@ -715,7 +711,7 @@ public abstract class TransactionManagers {
             return createRawLeaderServices(metricsManager, config.leader().get(), env, lock, time, userAgent);
         } else if (config.timestamp().isPresent() && config.lock().isPresent()) {
             return createRawRemoteServices(metricsManager, config, userAgent);
-        } else if (isUsingTimeLock(config, initialRuntimeConfig)) {
+        } else if (isUsingTimeLock(initialRuntimeConfig)) {
             return createRawServicesFromTimeLock(metricsManager, config, runtimeConfigSupplier, invalidator, userAgent);
         } else {
             return createRawEmbeddedServices(metricsManager, env, lock, time);
@@ -725,8 +721,6 @@ public abstract class TransactionManagers {
     private static void assertNoSpuriousTimeLockBlockInRuntimeConfig(
             AtlasDbConfig config,
             AtlasDbRuntimeConfig initialRuntimeConfig) {
-        // Note: The other direction (timelock install config without a runtime block) should be maintained for
-        // backwards compatibility.
         if (remoteTimestampAndLockOrLeaderBlocksPresent(config) && initialRuntimeConfig.timelockRuntime().isPresent()) {
             throw new IllegalStateException("Found a service configured not to use timelock, with a timelock"
                     + " block in the runtime config! This is unexpected. If you wish to use non-timelock services,"
@@ -761,10 +755,8 @@ public abstract class TransactionManagers {
             Supplier<AtlasDbRuntimeConfig> runtimeConfigSupplier) {
         Preconditions.checkState(!remoteTimestampAndLockOrLeaderBlocksPresent(config),
                 "Cannot create raw services from timelock with another source of timestamps/locks configured!");
-        TimeLockClientConfig clientConfig = config.timelock().orElse(ImmutableTimeLockClientConfig.builder().build());
-        String resolvedClient = OptionalResolver.resolve(clientConfig.client(), config.namespace());
-        return () -> ServerListConfigs.parseInstallAndRuntimeConfigs(
-                clientConfig,
+        String resolvedClient = config.namespace().get();
+        return () -> ServerListConfigs.parseRuntimeConfigs(
                 () -> runtimeConfigSupplier.get().timelockRuntime(),
                 resolvedClient);
     }