+++ docs/source/release_notes/release-notes.rst
@@ -50,6 +50,9 @@ develop
     *    - Type
          - Change
 
+    *    - |userbreak|
+         - The ``timelock`` block from the install config has been removed. The block was deprecated in v0.67.0 (released on 15th of November 2017).
+
     *    - |improved|
          - TimeLock Server now exposes a ``startAtlasDbTransaction`` endpoint which locks an immutable timestamp and then gets a fresh timestamp (in a single round-trip call); new TimeLock clients call this endpoint.
            This saves an estimated one TimeLock round-trip of latency when starting a transaction.